Move worker modules to separate folder.
