+++ common/models/client.js
@@ -1,6 +1,5 @@
 'use strict';
 const appRoot = require('app-root-path');
-const debugMockData = require('debug')('order-manager:Client:mockData');
 const debugMap = require('debug')('order-manager:Map:update');
 const Promise = require('bluebird');
 const yn = require('yn');
@@ -153,68 +152,6 @@ module.exports = function(Client) {
       result.error_message = err;
     }
     return result;
-  };
-
-  /**
-   * create 5 mock clients. Mock clients have an email 'mockClient@etr.com'
-   * @returns {Client[]} - array of mock clients
-   */
-  Client.mockData = async function() {
-    if (!yn(process.env.CREATE_MOCK_DATA)) {
-      return [];
-    }
-    debugMockData('Client.mockData() - Begins');
-    let clients = [];
-    const MockEmail = 'mockClient@etr.com';
-    clients = await Client.find({ where: { email: MockEmail } });
-    if (clients.length === 0) {
-      const MockClients = [
-        {
-          name: 'Kenny Market',
-          addressStreet: '3508 W 3rd St.', addressCity: 'Los Angeles', addressState: 'CA', addressZip: '90020',
-          phone: '213-555-1111', email: MockEmail, contactPersonName: 'Kenny T. Ro', contactPersonPhone: '213-555-1112',
-          feeType: 'Rate', feeValue: 3.50, showPublic: false
-        },
-        {
-          name: 'Jose Food',
-          addressStreet: '928 S Western Ave #257.', addressCity: 'Los Angeles', addressState: 'CA', addressZip: '90006',
-          phone: '213-555-2222', email: MockEmail, contactPersonName: 'Jose Ro', contactPersonPhone: '213-555-2221',
-          feeType: 'Rate', feeValue: 4.50, showPublic: false
-        },
-        {
-          name: 'LA Grocery',
-          addressStreet: '2915 W Olympic Blvd.', addressCity: 'Los Angeles', addressState: 'CA', addressZip: '90006',
-          phone: '213-555-3333', email: MockEmail, contactPersonName: 'Kenny La', contactPersonPhone: '213-555-3332',
-          feeType: 'Fixed', feeValue: 100.00, showPublic: false
-        },
-        {
-          name: 'Kim\'s Dollar Store',
-          addressStreet: '3917 W 6th St.', addressCity: 'Los Angeles', addressState: 'CA', addressZip: '90020',
-          phone: '213-555-4444', email: MockEmail, contactPersonName: 'Lan Y. Kim', contactPersonPhone: '213-555-4442',
-          feeType: 'Rate', feeValue: 4.00, showPublic: false
-        },
-        {
-          name: 'USA Market',
-          addressStreet: '1471 W Jefferson Blvd.', addressCity: 'Los Angeles', addressState: 'CA', addressZip: '90007',
-          phone: '213-555-5551', email: MockEmail, contactPersonName: 'Steve Trump', contactPersonPhone: '213-555-5552',
-          feeType: 'Rate', feeValue: 5.50, showPublic: false
-        }
-      ];
-      try {
-        const routes = await app.models.DeliveryRoute.find();
-        MockClients.forEach(function(client, index) {
-          client.deliveryRouteId = routes[index % routes.length].id;
-        });
-        clients = await Client.create(MockClients);
-        clients.forEach(function(client) {
-          debugMockData(`<Client[${client.id}]>: Created new client`);
-        });
-      } catch (error) {
-        console.error(`Failed to create mock clients - ${error.message}`);
-      }
-    }
-    debugMockData('Client.mockData() - Ends');
-    return clients;
   };
 
   /**
